fetch: split http module into error and auth
